refactor: move JwtStrategy to identity module
