+++ src/producer/producer.module.ts
@@ -11,7 +11,7 @@ import { ProducerController } from './web/rest/controller/producer.controller';
 import { EncryptPassword } from './domain/model/encrypt-password';
 import { PasswordFactory } from './domain/model/password.factory';
 import { JwtService } from '@nestjs/jwt';
-import { JwtStrategy } from '../shared/auth/strategy/jwt.strategy';
+import { JwtStrategy } from '../identity/auth-strategy/jwt.strategy';
 
 @Module({
   imports: [AppTypeOrmModule],
